+++ main/app/src/main/java/com/crazydiamond/seatassist/SeatAssistScreen.kt
@@ -1,0 +1,12 @@
+package com.crazydiamond.seatassist
+
+enum class SeatAssistScreen() {
+    Main(),
+    Members(),
+    MembersNumber(),
+    Lottery(),
+    Custom(),
+    Start(),
+    Usage(),
+    Splash()
+}
