--- main/app/src/main/java/com/example/seatassist/SeatAssistScreen.kt
@@ -1,12 +1,0 @@
-package com.example.seatassist
-
-enum class SeatAssistScreen() {
-    Main(),
-    Members(),
-    MembersNumber(),
-    Lottery(),
-    Custom(),
-    Start(),
-    Usage(),
-    Splash()
-}
