--- main/app/src/main/java/com/example/seatassist/ui/theme/Color.kt
@@ -1,12 +1,0 @@
-package com.example.seatassist.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Purple700 = Color(0xFF3700B3)
-val Teal200 = Color(0xFF03DAC5)
-
-val Sidecar = Color(0xFFF5EFBB)
-val PickledBlueWood = Color(0xFF324B54)
-val AkaRoa = Color(0xFFDBD2AC)
